Crate tm4c123x_hal

HAL for the tm4c123x family of microcontrollers

This is an implementation of the embedded-hal traits for the tm4c123x family of microcontrollers.

Usage

To build applications (binary crates) using this crate follow the cortex-m-quickstart instructions and add this crate as a dependency in step number 5 and make sure you enable the “rt” Cargo feature of this crate.

Examples

Examples of using these abstractions like these can be found in the documentation of the f3 crate.

Re-exports

pub use tm4c123x;

Modules

bb

Code to handle bit-banding.

delay

Code for busy-waiting

gpio

General Purpose Input / Output

hib

A wrapper for the HIB (Hibernation) peripheral

i2c

Inter-Integrated Circuit (I2C) bus

prelude

Prelude

pwm

PWM abstractions

serial

Serial

spi

Serial Peripheral Interface (SPI) bus

sysctl

System Control

time

Time units

timer

Timers

Structs

CorePeripherals

Core peripherals

Peripherals

All the peripherals