Struct tm4c123x_hal::gpio::gpiob::PB3

pub struct PB3<MODE> { /* private fields */ }

Pin

Implementations

impl<MODE> PB3<MODE> where
MODE: IsUnlocked,

pub fn into_af_push_pull<AF>(
self,
_gpio_control: &mut GpioControl
) -> PB3<AlternateFunction<AF, PushPull>> where
AF: AlternateFunctionChoice,

Configures the pin to serve as alternate function 1 through 15. Disables open-drain to make the output a push-pull.

pub fn into_af_pull_up<AF>(
self,
_gpio_control: &mut GpioControl
) -> PB3<AlternateFunction<AF, PullUp>> where
AF: AlternateFunctionChoice,

Configures the pin to serve as alternate function 1 through 15 with a weak pull-up resistor.

pub fn into_af_pull_down<AF>(
self,
_gpio_control: &mut GpioControl
) -> PB3<AlternateFunction<AF, PullDown>> where
AF: AlternateFunctionChoice,

Configures the pin to serve as alternate function 1 through 15 with a weak pull-down resistor.

pub fn into_af_open_drain<AF, ODM>(
self,
_gpio_control: &mut GpioControl
) -> PB3<AlternateFunction<AF, OpenDrain<ODM>>> where
AF: AlternateFunctionChoice,
ODM: OpenDrainMode,

Configures the pin to serve as alternate function 1 through 15. Enables open-drain (useful for I2C SDA, for example).

pub fn into_floating_input(self) -> PB3<Input<Floating>>

Configures the pin to operate as a floating input pin

pub fn into_pull_down_input(self) -> PB3<Input<PullDown>>

Configures the pin to operate as a pulled down input pin

pub fn into_pull_up_input(self) -> PB3<Input<PullUp>>

Configures the pin to operate as a pulled up input pin

pub fn into_open_drain_output<ODM>(self) -> PB3<Output<OpenDrain<ODM>>> where
ODM: OpenDrainMode,

Configures the pin to operate as an open drain output pin

pub fn into_push_pull_output(self) -> PB3<Output<PushPull>>

Configures the pin to operate as an push pull output pin

pub fn into_tri_state(self) -> PB3<Tristate>

Configures the pin as tri-state

impl<MODE> PB3<MODE>

pub fn downgrade(self) -> PBx<MODE>

Erases the pin number from the type

This is useful when you want to collect the pins into an array where you need all the elements to have the same type

impl<MODE> PB3<Input<MODE>> where
MODE: InputMode,

pub fn set_interrupt_mode(&mut self, mode: InterruptMode)

Enables or disables interrupts on this GPIO pin.

pub fn get_interrupt_status(&self) -> bool

Returns the current interrupt status for this pin.

pub fn clear_interrupt(&self)

Marks the interrupt for this pin as handled. You should call this (or perform its functionality) from the ISR.

impl PB3<Locked>

pub fn unlock(self, _gpio_control: &mut GpioControl) -> PB3<Tristate>

Unlock a GPIO so that it can be used. This is required on ‘special’ GPIOs that the manufacturer doesn’t want you to change by accident - like NMI and JTAG pins.