Struct tm4c123x_hal::spi::Spi

pub struct Spi<SPI, PINS> { /* private fields */ }

SPI peripheral operating in full duplex master mode

Implementations

impl<SCK, MISO, MOSI> Spi<SSI0, (SCK, MISO, MOSI)>

pub fn spi0<F>(
spi: SSI0,
pins: (SCK, MISO, MOSI),
mode: Mode,
freq: F,
clocks: &Clocks,
pc: &PowerControl
) -> Self where
F: Into<Hertz>,
SCK: SckPin<SSI0>,
MISO: MisoPin<SSI0>,
MOSI: MosiPin<SSI0>,

Configures the SPI peripheral to operate in full duplex master mode

pub fn free(self) -> (SSI0, (SCK, MISO, MOSI))

Releases the SPI peripheral and associated pins

pub fn reclock<F>(&mut self, freq: F, clocks: &Clocks) where
F: Into<Hertz>,

Change the clock frequency of the SPI device.

impl<SCK, MISO, MOSI> Spi<SSI1, (SCK, MISO, MOSI)>

pub fn spi1<F>(
spi: SSI1,
pins: (SCK, MISO, MOSI),
mode: Mode,
freq: F,
clocks: &Clocks,
pc: &PowerControl
) -> Self where
F: Into<Hertz>,
SCK: SckPin<SSI1>,
MISO: MisoPin<SSI1>,
MOSI: MosiPin<SSI1>,

Configures the SPI peripheral to operate in full duplex master mode

pub fn free(self) -> (SSI1, (SCK, MISO, MOSI))

Releases the SPI peripheral and associated pins

pub fn reclock<F>(&mut self, freq: F, clocks: &Clocks) where
F: Into<Hertz>,

Change the clock frequency of the SPI device.

impl<SCK, MISO, MOSI> Spi<SSI2, (SCK, MISO, MOSI)>

pub fn spi2<F>(
spi: SSI2,
pins: (SCK, MISO, MOSI),
mode: Mode,
freq: F,
clocks: &Clocks,
pc: &PowerControl
) -> Self where
F: Into<Hertz>,
SCK: SckPin<SSI2>,
MISO: MisoPin<SSI2>,
MOSI: MosiPin<SSI2>,

Configures the SPI peripheral to operate in full duplex master mode

pub fn free(self) -> (SSI2, (SCK, MISO, MOSI))

Releases the SPI peripheral and associated pins

pub fn reclock<F>(&mut self, freq: F, clocks: &Clocks) where
F: Into<Hertz>,

Change the clock frequency of the SPI device.

impl<SCK, MISO, MOSI> Spi<SSI3, (SCK, MISO, MOSI)>

pub fn spi3<F>(
spi: SSI3,
pins: (SCK, MISO, MOSI),
mode: Mode,
freq: F,
clocks: &Clocks,
pc: &PowerControl
) -> Self where
F: Into<Hertz>,
SCK: SckPin<SSI3>,
MISO: MisoPin<SSI3>,
MOSI: MosiPin<SSI3>,

Configures the SPI peripheral to operate in full duplex master mode

pub fn free(self) -> (SSI3, (SCK, MISO, MOSI))

Releases the SPI peripheral and associated pins

pub fn reclock<F>(&mut self, freq: F, clocks: &Clocks) where
F: Into<Hertz>,

Change the clock frequency of the SPI device.